+++ CHANGELOG.md
@@ -99,6 +99,12 @@ duplication."
 
 ## Unreleased
 
+### Unreleased: Generic Generator
+
+- The workflow that calls the provenance generator re-usable workflows
+  will fail if an adjacent Jobs in the calling workflow are using self-hosted runners.
+
+
 ### Unreleased: Gradle Builder
 
 - The Gradle Builder was fixed when the project root is the same as the